Changed entrance to web app to land at a welcome page with a short blurb about the project
Main change: proper folder structure of HTML documents. Now, index.html is no longer the landing page and web app in one.
The index.html page contains the welcome page with a short blurb about the project and a link to the web app.
Controller functionality has been changed to reflect this change.